Revert "Add health-check to amqp"
