+++ src/healthcheck.js
@@ -12,7 +12,6 @@ function getService(value) {
   let url;
   let route;
   let knex;
-  let queue;
   switch (typeof value) {
     case 'string':
       url = value;
@@ -26,11 +25,6 @@ function getService(value) {
         break;
       }
 
-      if (value.queue) {
-        queue = value.queue;
-        break;
-      }
-
       url = value.url;
       route = value.route;
       break;
@@ -38,7 +32,7 @@ function getService(value) {
       return {};
   }
 
-  return { url, route, knex, queue };
+  return { url, route, knex };
 }
 
 /**
