+++ test/database.js
@@ -1,18 +1,31 @@
 import { feature } from 'ava-spec';
+import knex from 'knex';
+import mockKnex from 'mock-knex';
 
 import database from '../lib/database';
-import {
-  knexUp,
-  knexDown,
-  mockKnex,
-  knexTracker,
-} from './consts';
+
+const knexUp = knex({
+  client: 'mysql',
+  connection: {
+    host: 'localhost',
+    user: 'user',
+    password: 'password',
+    database: 'database',
+    port: 3306,
+  },
+  debug: false,
+});
+
+const knexDown = knex({
+  client: 'mysql',
+});
 
 feature('testing database connection', scenario => {
   scenario('given that the database is up', async t => {
+    const tracker = mockKnex.getTracker();
     mockKnex.mock(knexUp);
-    knexTracker.install();
-    knexTracker.on('query', query => {
+    tracker.install();
+    tracker.on('query', query => {
       query.response();
     });
